applications: serial_lte_modem: Refactor power and indicate pin handling
Move power and indicate pin handling into an own module from main.c.
