+++ applications/serial_lte_modem/src/main.c
@@ -4,28 +4,19 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
 #include <zephyr/logging/log.h>
-#include <zephyr/logging/log_ctrl.h>
 #include <zephyr/kernel.h>
-#include <assert.h>
-#include <stdio.h>
-#include <zephyr/drivers/uart.h>
-#include <zephyr/drivers/gpio.h>
-#include <string.h>
 #include <nrf_modem.h>
-#include <hal/nrf_gpio.h>
 #include <hal/nrf_power.h>
-#include <hal/nrf_regulators.h>
 #include <modem/nrf_modem_lib.h>
 #include <zephyr/dfu/mcuboot.h>
 #include <dfu/dfu_target.h>
 #include <zephyr/sys/reboot.h>
-#include <zephyr/drivers/clock_control.h>
-#include <zephyr/drivers/clock_control/nrf_clock_control.h>
 #include <net/fota_download.h>
 #include "slm_at_host.h"
 #include "slm_at_fota.h"
 #include "slm_settings.h"
 #include "slm_util.h"
+#include "slm_ctrl_pin.h"
 
 LOG_MODULE_REGISTER(slm, CONFIG_SLM_LOG_LEVEL);
 
@@ -33,22 +24,7 @@ LOG_MODULE_REGISTER(slm, CONFIG_SLM_LOG_LEVEL);
 #define SLM_WQ_PRIORITY		K_LOWEST_APPLICATION_THREAD_PRIO
 static K_THREAD_STACK_DEFINE(slm_wq_stack_area, SLM_WQ_STACK_SIZE);
 
-static const struct device *gpio_dev = DEVICE_DT_GET(DT_NODELABEL(gpio0));
-#if POWER_PIN_IS_ENABLED
-static struct gpio_callback gpio_cb;
-#else
-BUILD_ASSERT(!IS_ENABLED(CONFIG_SLM_START_SLEEP),
-	"CONFIG_SLM_START_SLEEP requires CONFIG_SLM_POWER_PIN to be defined.");
-#endif
-static struct k_work_delayable indicate_work;
-
 struct k_work_q slm_work_q;
-static atomic_t callback_wakeup_running;
-
-/* Forward declarations */
-static void indicate_wk(struct k_work *work);
-static void power_pin_callback_wakeup(const struct device *dev,
-				      struct gpio_callback *gpio_callback, uint32_t);
 
 NRF_MODEM_LIB_ON_INIT(lwm2m_init_hook, on_modem_lib_init, NULL);
 NRF_MODEM_LIB_ON_DFU_RES(main_dfu_hook, on_modem_dfu_res, NULL);
@@ -88,276 +64,6 @@ void nrf_modem_fault_handler(struct nrf_modem_fault_info *fault_info)
 	k_work_submit(&modem_failure_work);
 }
 #endif /* CONFIG_NRF_MODEM_LIB_ON_FAULT_APPLICATION_SPECIFIC */
-
-static int ext_xtal_control(bool xtal_on)
-{
-	int err = 0;
-#if defined(CONFIG_SLM_EXTERNAL_XTAL)
-	static struct onoff_manager *clk_mgr;
-
-	if (xtal_on) {
-		struct onoff_client cli = {};
-
-		/* request external XTAL for UART */
-		clk_mgr = z_nrf_clock_control_get_onoff(CLOCK_CONTROL_NRF_SUBSYS_HF);
-		sys_notify_init_spinwait(&cli.notify);
-		err = onoff_request(clk_mgr, &cli);
-		if (err < 0) {
-			LOG_ERR("Clock request failed: %d", err);
-			return err;
-		}
-		while (sys_notify_fetch_result(&cli.notify, &err) < 0) {
-			/*empty*/
-		}
-	} else {
-		/* release external XTAL for UART */
-		err = onoff_release(clk_mgr);
-		if (err < 0) {
-			LOG_ERR("Clock release failed: %d", err);
-			return err;
-		}
-	}
-#endif
-
-	return err;
-}
-
-#if POWER_PIN_IS_ENABLED
-
-static int configure_power_pin_interrupt(gpio_callback_handler_t handler, gpio_flags_t flags)
-{
-	int err;
-	const gpio_pin_t pin = CONFIG_SLM_POWER_PIN;
-
-	/* First disable the previously configured interrupt. Somehow when in idle mode if
-	 * the wake-up interrupt is configured to be on an edge the power consumption
-	 * drastically increases (3x), which is is why it is configured to be level-triggered.
-	 * When entering idle for some reason first disabling the previously edge-level
-	 * configured interrupt is also needed to keep the power consumption down.
-	 */
-	gpio_pin_interrupt_configure(gpio_dev, pin, GPIO_INT_DISABLE);
-
-	err = gpio_pin_interrupt_configure(gpio_dev, pin, flags);
-	if (err) {
-		LOG_ERR("Failed to configure %s (0x%x) on power pin. (%d)",
-			"interrupt", flags, err);
-		return err;
-	}
-
-	gpio_init_callback(&gpio_cb, handler, BIT(pin));
-
-	err = gpio_add_callback(gpio_dev, &gpio_cb);
-	if (err) {
-		LOG_ERR("Failed to configure %s (0x%x) on power pin. (%d)", "callback", flags, err);
-		return err;
-	}
-
-	LOG_DBG("Configured interrupt (0x%x) on power pin (%u) with handler (%p).",
-		flags, pin, handler);
-	return 0;
-}
-
-static void slm_enter_sleep_work_fn(struct k_work *)
-{
-	slm_enter_sleep();
-}
-
-static void power_pin_callback_poweroff(const struct device *dev,
-					struct gpio_callback *gpio_callback, uint32_t)
-{
-	static K_WORK_DEFINE(work, slm_enter_sleep_work_fn);
-
-	LOG_INF("Power off triggered.");
-	gpio_remove_callback(dev, gpio_callback);
-	k_work_submit(&work);
-}
-
-#endif /* POWER_PIN_IS_ENABLED */
-
-#if POWER_PIN_IS_ENABLED || INDICATE_PIN_IS_ENABLED
-
-static int configure_gpio(gpio_pin_t pin, gpio_flags_t flags)
-{
-	const int err = gpio_pin_configure(gpio_dev, pin, flags);
-
-	if (err) {
-		LOG_ERR("Failed to configure GPIO pin P0.%d. (%d)", pin, err);
-		return err;
-	}
-
-	return 0;
-}
-#endif
-
-static int init_gpios(void)
-{
-	if (!device_is_ready(gpio_dev)) {
-		LOG_ERR("GPIO controller not ready");
-		return -ENODEV;
-	}
-
-#if POWER_PIN_IS_ENABLED
-	if (configure_gpio(CONFIG_SLM_POWER_PIN, GPIO_INPUT | GPIO_PULL_UP | GPIO_ACTIVE_LOW)) {
-		return -1;
-	}
-#endif
-
-#if INDICATE_PIN_IS_ENABLED
-	if (configure_gpio(CONFIG_SLM_INDICATE_PIN, GPIO_OUTPUT_INACTIVE | GPIO_ACTIVE_LOW)) {
-		return -1;
-	}
-#endif
-
-	return 0;
-}
-
-int slm_indicate(void)
-{
-	int err = 0;
-
-#if (INDICATE_PIN_IS_ENABLED)
-	if (k_work_delayable_is_pending(&indicate_work)) {
-		return 0;
-	}
-	LOG_DBG("Start indicating");
-	err = gpio_pin_set(gpio_dev, CONFIG_SLM_INDICATE_PIN, 1);
-	if (err) {
-		LOG_ERR("GPIO_0 set error: %d", err);
-	} else {
-		k_work_reschedule(&indicate_work, K_MSEC(CONFIG_SLM_INDICATE_TIME));
-	}
-#endif
-	return err;
-}
-
-static void indicate_stop(void)
-{
-#if (INDICATE_PIN_IS_ENABLED)
-	if (gpio_pin_set(gpio_dev, CONFIG_SLM_INDICATE_PIN, 0) != 0) {
-		LOG_WRN("GPIO_0 set error");
-	}
-	LOG_DBG("Stop indicating");
-#endif
-}
-
-#if POWER_PIN_IS_ENABLED
-
-static void power_pin_callback_enable_poweroff(const struct device *dev,
-					       struct gpio_callback *gpio_callback, uint32_t)
-{
-	LOG_DBG("Enabling the poweroff interrupt shortly...");
-	gpio_remove_callback(dev, gpio_callback);
-
-	configure_power_pin_interrupt(power_pin_callback_poweroff, GPIO_INT_EDGE_RISING);
-}
-
-static void power_pin_callback_wakeup_work_fn(struct k_work *)
-{
-	int err;
-
-	LOG_INF("Resuming from idle.");
-	if (k_work_delayable_is_pending(&indicate_work)) {
-		k_work_cancel_delayable(&indicate_work);
-		indicate_stop();
-	}
-
-	err = ext_xtal_control(true);
-	if (err < 0) {
-		LOG_WRN("Failed to enable ext XTAL: %d", err);
-	}
-	err = slm_at_host_power_on();
-	if (err) {
-		LOG_ERR("Failed to power on uart: %d. Resetting SLM.", err);
-		gpio_remove_callback(gpio_dev, &gpio_cb);
-		slm_reset();
-		return;
-	}
-
-	atomic_set(&callback_wakeup_running, false);
-}
-
-static void power_pin_callback_wakeup(const struct device *dev,
-				      struct gpio_callback *gpio_callback, uint32_t)
-{
-	static K_WORK_DEFINE(work, power_pin_callback_wakeup_work_fn);
-
-	/* Prevent level triggered interrupt running this multiple times. */
-	if (!atomic_cas(&callback_wakeup_running, false, true)) {
-		return;
-	}
-
-	LOG_INF("Resuming from idle shortly...");
-	gpio_remove_callback(dev, gpio_callback);
-
-	/* Enable the poweroff interrupt only when the pin will be back to a nonactive state. */
-	configure_power_pin_interrupt(power_pin_callback_enable_poweroff, GPIO_INT_EDGE_RISING);
-
-	k_work_submit(&work);
-}
-
-void slm_enter_idle(void)
-{
-	LOG_INF("Entering idle.");
-	int err;
-
-	gpio_remove_callback(gpio_dev, &gpio_cb);
-
-	err = configure_power_pin_interrupt(power_pin_callback_wakeup, GPIO_INT_LEVEL_LOW);
-	if (err) {
-		return;
-	}
-
-	err = ext_xtal_control(false);
-	if (err < 0) {
-		LOG_WRN("Failed to disable ext XTAL: %d", err);
-	}
-}
-
-static void enter_sleep_no_uninit(void)
-{
-	LOG_INF("Entering sleep.");
-	LOG_PANIC();
-	nrf_gpio_cfg_sense_set(CONFIG_SLM_POWER_PIN, NRF_GPIO_PIN_SENSE_LOW);
-
-	k_sleep(K_MSEC(100));
-
-	nrf_regulators_system_off(NRF_REGULATORS_NS);
-	assert(false);
-}
-
-void slm_enter_sleep(void)
-{
-	slm_at_host_uninit();
-
-	/* Only power off the modem if it has not been put
-	 * in flight mode to allow reducing NVM wear.
-	 */
-	if (!slm_is_modem_functional_mode(LTE_LC_FUNC_MODE_OFFLINE)) {
-		slm_power_off_modem();
-	}
-	enter_sleep_no_uninit();
-}
-
-#endif /* POWER_PIN_IS_ENABLED */
-
-void slm_enter_shutdown(void)
-{
-	LOG_INF("Entering shutdown.");
-
-	/* De-configure GPIOs */
-#if POWER_PIN_IS_ENABLED
-	gpio_pin_interrupt_configure(gpio_dev, CONFIG_SLM_POWER_PIN, GPIO_INT_DISABLE);
-	gpio_pin_configure(gpio_dev, CONFIG_SLM_POWER_PIN, GPIO_DISCONNECTED);
-#endif
-#if INDICATE_PIN_IS_ENABLED
-	gpio_pin_configure(gpio_dev, CONFIG_SLM_INDICATE_PIN, GPIO_DISCONNECTED);
-#endif
-
-	k_sleep(K_MSEC(100));
-
-	nrf_regulators_system_off(NRF_REGULATORS_NS);
-	assert(false);
-}
 
 static void on_modem_dfu_res(int dfu_res, void *ctx)
 {
@@ -495,22 +201,9 @@ int start_execute(void)
 	int err;
 
 	LOG_INF("Serial LTE Modem");
-	err = ext_xtal_control(true);
-	if (err < 0) {
-		LOG_ERR("Failed to enable ext XTAL: %d", err);
-		return err;
-	}
-#if POWER_PIN_IS_ENABLED
-	/* Do not directly enable the poweroff interrupt so that only a full toggle triggers
-	 * power off. This is because power on is triggered on low level, so if the pin is held
-	 * down until SLM is fully initialized releasing it would directly trigger the power off.
-	 */
-	err = configure_power_pin_interrupt(power_pin_callback_enable_poweroff,
-					    GPIO_INT_EDGE_FALLING);
-	if (err) {
-		return err;
-	}
-#endif
+
+	slm_ctrl_pin_init();
+
 	/* This will send "READY" or "INIT ERROR" to UART so after this nothing
 	 * should be done that can fail
 	 */
@@ -528,24 +221,15 @@ int start_execute(void)
 	return 0;
 }
 
-static void indicate_wk(struct k_work *work)
-{
-	ARG_UNUSED(work);
-
-	indicate_stop();
-}
-
 int main(void)
 {
 	const uint32_t rr = nrf_power_resetreas_get(NRF_POWER_NS);
 
 	nrf_power_resetreas_clear(NRF_POWER_NS, 0x70017);
 	LOG_DBG("RR: 0x%08x", rr);
-	k_work_init_delayable(&indicate_work, indicate_wk);
 
-	if (init_gpios() != 0) {
-		LOG_WRN("Failed to init GPIO pins.");
-	}
+	slm_ctrl_pin_init_gpios();
+
 	/* Init and load settings */
 	if (slm_settings_init() != 0) {
 		LOG_WRN("Failed to init slm settings");
